Moved logictree.py in hazardlib
