+++ openquake/commonlib/__init__.py
@@ -18,3 +18,4 @@
 
 # keep the following line to have a __version__
 from openquake.baselib import __version__
+from openquake.hazardlib import logictree
