+++ openquake/commonlib/tests/source_test.py
@@ -28,9 +28,9 @@ from openquake.baselib.parallel import Starmap
 from openquake.hazardlib import site, geo, mfd, pmf, scalerel, tests as htests
 from openquake.hazardlib import source, sourceconverter as s
 from openquake.hazardlib.tom import PoissonTOM
-from openquake.commonlib import tests, readinput
-from openquake.commonlib.logictree import FullLogicTree
+from openquake.hazardlib.logictree import FullLogicTree
 from openquake.hazardlib import nrml
+from openquake.commonlib import tests, readinput
 
 # directory where the example files are
 NRML_DIR = os.path.dirname(htests.__file__)
